refac: create jwt pkg #5
